Refactor for apt command line tools
