+++ .gitignore
@@ -6,6 +6,8 @@ bin/*
 export/*
 gulibrary/*
 core/target/*
+core-utils/target/*
+bag-utils/target/*
 dspace/target/*
 demo/target/*
 gulibrary/target/*
