+++ bag-utils/src/main/edu/georgetown/library/fileAnalyzer/util/ZipBagHelper.java
