+++ dspace/pom.xml
@@ -21,7 +21,11 @@
 	</properties>
 
 	<dependencies>
-
+        <dependency>
+            <groupId>edu.georgetown.library</groupId>
+            <artifactId>FileAnalyzerCoreUtils</artifactId>
+            <version>2.0</version>
+        </dependency>
 		<dependency>
 			<groupId>edu.georgetown.library</groupId>
 			<artifactId>CoreFileAnalyzer</artifactId>
