+++ dspace/src/main/edu/georgetown/library/fileAnalyzer/importer/DSpaceMetadata2Marc.java
@@ -31,7 +31,7 @@ import gov.nara.nwts.ftapp.stats.StatsGenerator;
 import gov.nara.nwts.ftapp.stats.StatsItem;
 import gov.nara.nwts.ftapp.stats.StatsItemConfig;
 import gov.nara.nwts.ftapp.stats.StatsItemEnum;
-import gov.nara.nwts.ftapp.util.XMLUtil;
+import edu.georgetown.library.fileAnalyzer.util.XMLUtil;
 
 /**
  * Importer for tab delimited files
