+++ dspace/src/main/edu/georgetown/library/fileAnalyzer/importer/IngestFolderCreate.java
@@ -6,6 +6,7 @@ import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.FilenameFilter;
 import java.io.IOException;
+import java.nio.file.Files;
 import java.text.DateFormat;
 import java.text.NumberFormat;
 import java.text.SimpleDateFormat;
@@ -38,9 +39,8 @@ import gov.nara.nwts.ftapp.stats.StatsGenerator;
 import gov.nara.nwts.ftapp.stats.StatsItem;
 import gov.nara.nwts.ftapp.stats.StatsItemConfig;
 import gov.nara.nwts.ftapp.stats.StatsItemEnum;
-import gov.nara.nwts.ftapp.util.FileUtil;
-import gov.nara.nwts.ftapp.util.XMLUtil;
-import gov.nara.nwts.ftapp.util.ZipUtil;
+import edu.georgetown.library.fileAnalyzer.util.XMLUtil;
+import edu.georgetown.library.fileAnalyzer.util.ZipUtil;
 
 /**
  * @author TBrady
@@ -582,7 +582,7 @@ public class IngestFolderCreate extends DefaultImporter {
 				source.renameTo(dest);
 				stats.setVal(sienum, FileStats.MOVED_TO_INGEST);
 			} else if (mode == MODE.COPY) {
-				FileUtil.copyFile(source, dest);
+				Files.copy(source.toPath(), dest.toPath());
 				stats.setVal(sienum, FileStats.COPIED_TO_INGEST);
 			}
